Golf

Exercise 1

skips :: [a] -> [[a]]

skips xs takes a list xs and returns a new list where every ith item is a list formed by keeping every ith element from the original xs.

skipN :: Integral b => [a] -> b -> [a]

skipN xs n takes a list xs and returns a new list where every nth item is kept. It works by first assigning an index to each value by creating a tuple of the form (i,x) where i is the index and x the value. Then to determine whether or not to keep that ith item, we check that the value i + n is divisible by n.

>>> skipN "abcd" 1
"abcd"

>>> skipN "abcd" 2
"bd"

>>> skipN "abcd" 3
"c"

Exercise 2

localMaxima :: [Integer] -> [Integer]

localMaxima takes a list and returns a list of its local maxima, defined as an element that is strictly greater than both the previous and next elements.

Algorithm example: [2,9,5,6,1]

Consider a list of 3-element sequences from the given list

[(2,9,5),(9,5,6),(5,6,1)]

Filter this list, keeping triples that have a local maximum

[(2,9,5),(5,6,1)]

Extract the 2nd element from each of these triples

[9,6]

groupBy3 :: [a] -> [(a, a, a)]

isLocalMaximum :: Ord a => (a, a, a) -> Bool

snd3 :: (a, b, c) -> b

Exercise 3

histogram :: [Int] -> String

histogram takes a list of integers 0-9 and returns a textual representation of a vertical histogram, with a star per ocurrance of the digit.

Algorithm:

Ex. given [1,1,5,6]

Make a horizontal histogram

Count the number of ocurrances of each digit in a new list with corresponding indicies

[0,2,0,0,0,1,1,0,0,0]

Replace these counts with a string representation of "*"s

[""," * *","","",""," *"," *","","",""]

Pad with whitespace so we get a square matrix

      ["    "," * *","    ","    ","    "," *  "," *  ","    ","    ","    "]
         ^      ^      ^      ^      ^      ^      ^      ^      ^      ^
         0      1      2      3      4      5      6      7      8      9
      

Add labels to each of these

      ["0=    ","1= * *","2=    ","3=    ","4=    "
      ,"5= *  ","6= *  ","7=    ","8=    ","9=    "
      ]
      

Transform horizontal histogram to vertical

horizontalHistogram :: [Int] -> [String]

count :: Eq a => a -> [a] -> Int

Count the occurances of x in a list

counts :: Eq a => [a] -> [a] -> [Int]

Similar to count, but instead takes a list of items to search for. The result is a list of counts, with indecies corresponding to the index of elementes from countThese

starify :: Int -> String

addLabels :: [String] -> [String]

makeSquare :: [String] -> [String]